update: better code by clippy

+++ src/ray.rs
@@ -13,8 +13,8 @@ impl Ray{
         assert!(scene.width > scene.height);
         let fov_adjustment = f64::from(scene.fov.to_radians() / 2.0).tan();
         let aspect_ratio = f64::from(scene.width) / f64::from(scene.height);
-        let sensor_x = ((f64::from(x) + 0.5) / f64::from(scene.width) * 2.0 - 1.0) * aspect_ratio * f64::from(fov_adjustment);
-        let sensor_y = (1.0 - (f64::from(y) + 0.5) / f64::from(scene.height) * 2.0) * f64::from(fov_adjustment);
+        let sensor_x = ((f64::from(x) + 0.5) / f64::from(scene.width) * 2.0 - 1.0) * aspect_ratio * fov_adjustment;
+        let sensor_y = (1.0 - (f64::from(y) + 0.5) / f64::from(scene.height) * 2.0) * fov_adjustment;
 
         Ray{
             origin: Vector3::zero(),
